requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-asgi-testclient==1.4.11
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.0
black==24.8.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.0
fastapi-cache2==0.2.2
fastapi-cli==0.0.5
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
orjson==3.10.6
packaging==24.1
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.3.3
pluggy==1.5.0
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==4.6.0
requests==2.32.3
rich==13.7.1
ruff==0.5.7
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
starlette==0.37.2
time-machine==2.14.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
